emerald run -d will kill the PID in the file only if it is not referring the current process

+++ app/terminal/main.js
@@ -73,8 +73,11 @@ program
                     var pid = parseInt(_pid, 10);
 
                     try {
-                        process.kill(pid, 'SIGKILL');
+                        if (pid !== process.pid) {
+                            process.kill(pid, 'SIGKILL');
+                        }
                     } catch (e){
+
                     }
 
                     callback(null);
